feat: getにhandlerを登録

+++ server/domain/src/repository/form_repository.rs
@@ -7,5 +7,5 @@ use crate::form::models::{Form, FormId, FormTitle};
 #[async_trait]
 pub trait FormRepository: Send + Sync + 'static {
     async fn create(&self, title: FormTitle) -> anyhow::Result<FormId>;
-    async fn list(&self, offset: i32, limit: i32) -> anyhow::Result<Vec<Form>>;
+    async fn list(&self, offset: i64, limit: i64) -> anyhow::Result<Vec<Form>>;
 }
