+++ server/infra/resource/src/repository/form_repository_impl.rs
@@ -15,7 +15,7 @@ impl<Client: DatabaseComponents + 'static> FormRepository for Repository<Client>
         self.client.form().create(title).await
     }
 
-    async fn list(&self, offset: i32, limit: i32) -> anyhow::Result<Vec<Form>> {
+    async fn list(&self, offset: i64, limit: i64) -> anyhow::Result<Vec<Form>> {
         self.client.form().list(offset, limit).await
     }
 }
